requirements.txt
fastapi
uvicorn
python-dotenv
openai
pdfminer.six
pdfplumber
streamlit
requests
python-multipart